+++ hardware/bean/avr/cores/bean/BeanSerialTransport.h
@@ -47,7 +47,7 @@ class BeanSerialTransport : public HardwareSerial {
   void BTSetAdvertisingInterval(uint16_t interval_ms);
   void BTSetConnectionInterval(const int interval_ms);
   void BTSetLocalName(const char *name);
-  void BTSetPairingPin(const uint16_t pin);
+  void BTSetPairingPin(const uint32_t pin);
   void BTSetTxPower(const BT_TXPOWER_DB_T &power);
   void BTSetScratchChar(BT_SCRATCH_T *setting, uint8_t length);
   int BTGetScratchChar(uint8_t scratchNum, ScratchData *scratchData);
